adding thread apis/hooks
